Module Odoc_odoc.Html_fragment

Produces html fragment files from a mld file.

val from_mld : xref_base_uri:string -> resolver:Resolver.t -> output:Fs.File.t -> warnings_options:Odoc_model.Error.warnings_options -> Fs.File.t -> (unit, [> Or_error.msg ]) Or_error.result

from_mld ~xref_base_uri ~resolver ~output input parses the content of the input file as a documentation page (i.e. the ocamldoc syntax), generates the equivalent HTML representation and writes the result into the output file. The produced file is an HTML fragment that can be embedded into other documents.

Cross-reference resolution uses the provided xref_base_uri to locate docset packages.