Module Odoc_odoc.Odoc_link

val link_page : resolver:Resolver.t -> filename:string -> Odoc_model.Lang.Page.t -> Odoc_model.Lang.Page.t Odoc_model.Error.with_warnings

val link_impl : resolver:Resolver.t -> filename:string -> Odoc_model.Lang.Implementation.t -> Odoc_model.Lang.Implementation.t Odoc_model.Error.with_warnings

val content_for_hidden_modules : Odoc_model.Lang.Signature.item list

val link_unit : resolver:Resolver.t -> filename:string -> Odoc_model.Lang.Compilation_unit.t -> Odoc_model.Lang.Compilation_unit.t Odoc_model.Error.with_warnings

val handle_warnings : input_warnings:Odoc_model.Error.t list -> warnings_options:Odoc_model.Error.warnings_options -> 'a Odoc_model.Error.with_warnings -> ('b * Odoc_model.Error.t list, [> `Msg of string ]) Or_error.result

~input_warnings are the warnings stored in the input file

val from_odoc : resolver:Resolver.t -> warnings_options:Odoc_model.Error.warnings_options -> Fs.File.t -> Fs.File.t -> ([> `Impl of Odoc_model.Lang.Implementation.t | `Module of Odoc_model.Lang.Compilation_unit.t | `Page of Odoc_model.Lang.Page.t | `Source_tree of Odoc_model.Lang.SourceTree.t ], [> Or_error.msg ]) Or_error.result

Read the input file and write to the output file. Also return the resulting tree.